+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-08-27 22:12:18
-// Commit: feat(perf): integrate custom telemetry logger in perf
+// Updated at: 2026-08-28 17:06:57
+// Commit: perf(perf): optimize memory consumption in perf list viewport
 
 class perfComponentState {
-  final String commitRef = '3976';
-  final String lastMessage = 'feat(perf): integrate custom telemetry logger in perf';
-  final DateTime timestamp = DateTime.parse('2026-08-27T22:12:18+05:00');
+  final String commitRef = '3993';
+  final String lastMessage = 'perf(perf): optimize memory consumption in perf list viewport';
+  final DateTime timestamp = DateTime.parse('2026-08-28T17:06:57+05:00');
 }